added viewmodel to fragment
