--- app/src/main/java/com/example/dadjoke1/BlankFragment.kt
@@ -1,25 +1,0 @@
-package com.example.dadjoke1
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-
-
-
-class BlankFragment : Fragment() {
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_blank, container, false)
-    }
-
-    companion object {
-        @JvmStatic
-        fun newInstance() = BlankFragment()
-    }
-}
