+++ app/src/main/java/com/example/dadjoke1/BlankFragment2.kt
@@ -1,0 +1,30 @@
+package com.example.dadjoke1
+
+import androidx.lifecycle.ViewModelProvider
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+
+class BlankFragment : Fragment() {
+
+    companion object {
+        fun newInstance() = BlankFragment()
+    }
+
+    private lateinit var viewModel: FragmentViewModel
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return inflater.inflate(R.layout.fragment_blank2, container, false)
+    }
+
+    override fun onActivityCreated(savedInstanceState: Bundle?) {
+        super.onActivityCreated(savedInstanceState)
+        viewModel = ViewModelProvider(this).get(FragmentViewModel::class.java)
+    }
+
+}
