+++ app/src/main/java/com/example/dadjoke1/FragmentViewModel.kt
@@ -1,0 +1,7 @@
+package com.example.dadjoke1
+
+import androidx.lifecycle.ViewModel
+
+class FragmentViewModel : ViewModel() {
+
+}
